fix: refactorizado el código moviendo el componente RequestCardsContainer a otro .jsx
