+++ frontend/src/components/RequestCardsContainer.jsx
@@ -1,0 +1,151 @@
+import { Box, Button, Card, CardContent, CardMedia, Typography, Tooltip, CardActions } from "@mui/material";
+
+const RequestCardsContainer = ({ requests, handleResponse, openConfirmModal, sent = false, isOwner= true }) => {
+    return (
+        <Box
+            sx={{
+                display: "flex",
+                flexDirection: "column",
+                gap: 2,
+                p: 2,
+                width: "100%",
+                maxWidth: "800px",
+                maxHeight: "75vh",
+                overflowY: "auto",
+                overflowX: "hidden",
+            }}
+        >
+        {requests.map((request) => (   
+            <Card
+                key={request.id}
+                sx={{
+                    display: "flex",
+                    flexDirection: "row",
+                    alignItems: "center",
+                    boxShadow: 3,
+                    p: 2,
+                    width: "100%",
+                    maxWidth: "750px",
+                    minHeight: "150px",
+                    borderRadius: 2,
+                    overflow: "hidden",
+                }}
+            >
+                <CardMedia
+                    component="img"
+                    sx={{
+                        width: 150,
+                        height: 150,
+                        objectFit: "cover",
+                        borderRadius: "2px",
+                        mr: 2,
+                        boxShadow: 1,
+                    }}
+                    image={request.imageUrl}
+                    alt={request.title}
+                />
+                <CardContent sx={{ flex: 1 }}>
+                    <Typography variant="h6" sx={{ fontWeight: "bold", mb: 1 }} >
+                        <a
+                            href={`show-item/${request.item.id}`}
+                            style={{ textDecoration: "none", color: "inherit" }}
+                        >
+                            {request.item.title}
+                        </a>
+                    </Typography>
+                    {isOwner && (
+                        <Typography variant="body2" color="textSecondary" sx={{ mb: 1 }}>
+                            <strong> Solicitado por: </strong>
+                            <Tooltip
+                                title={
+                                    <Card sx={{ width: 250 }}>
+                                        <CardContent>
+                                            <Typography variant="body2">
+                                                <strong>Nombre:</strong> {request.renter.name} {request.renter.surname}
+                                            </Typography>
+                                            <Typography variant="body2">
+                                                <strong>Email:</strong> {request.renter.email}
+                                            </Typography>
+                                        </CardContent>
+                                        <CardActions sx={{ justifyContent: "flex-end" }}>
+                                            {/* Botón para enviar mensaje al usuario, TODO implementar el chat */}
+                                            <Button size="small" onClick={() => alert("Enviando mensaje...")}>
+                                                Enviar Mensaje
+                                            </Button>
+                                        </CardActions>
+                                    </Card>
+                                }
+                                arrow
+                            >
+                                <a
+                                    href={`/perfil/${request.renter.username}`}
+                                    style={{
+                                        textDecoration: "none",
+                                        color: "#1976d2",
+                                        fontWeight: "bold",
+                                    }}
+                                >
+                                    {request.renter.name} {request.renter.surname}
+                                    {/* TODO añadir en el modelo de la renta el campo owner para poder serializarlo
+                                    y poner "solicitado a", de momento se deja así */}
+                                </a>
+                            </Tooltip>
+                        </Typography>
+                    )}
+                    <Typography variant="body2" sx={{ mb: 1 }}>
+                        <strong> Inicio: </strong> {new Date(request.start_date).toLocaleString('es-ES', {
+                            weekday: 'long',
+                            year: 'numeric',
+                            month: 'long',
+                            day: 'numeric',
+                            hour: '2-digit',
+                            minute: '2-digit'
+                        })}
+                    </Typography>
+                    <Typography variant="body2" sx={{ mb: 1 }}>
+                        <strong> Fin: </strong> {new Date(request.end_date).toLocaleString('es-ES', {
+                            weekday: 'long',
+                            year: 'numeric',
+                            month: 'long',
+                            day: 'numeric',
+                            hour: '2-digit',
+                            minute: '2-digit'
+                        })}
+                    </Typography>
+                    <Box sx={{ display: "flex", justifyContent: "flex-start", gap: 2 }}>
+                        {/* Si el estatus es 'aceptado' y el pago está pendiente, mostramos el botón de pago */}
+                        {request.rent_status === "accepted" && request.payment_status === "pending" && (
+                            <Button
+                                variant="contained"
+                                color="primary"
+                                size="small"
+                                onClick={() => alert("Iniciando proceso de pago...")} //TODO: Implementar el pago con Stripe
+                            >
+                                Pagar
+                            </Button>
+                        )}
+                        { isOwner && request.rent_status === "requested" && (<Button
+                            variant="contained"
+                            color="success"
+                            size="small"
+                            onClick={() => openConfirmModal(request, "accepted")}
+                        >
+                            Aceptar
+                        </Button>) }
+                        { isOwner && request.rent_status === "requested" && <Button
+                            variant="contained"
+                            color="error"
+                            size="small"
+                            onClick={() => openConfirmModal(request, "rejected")}
+                        >
+                            Rechazar
+                        </Button> }
+                    </Box>
+                </CardContent>
+            </Card>
+        ))}
+        </Box>
+    );
+};
+
+export default RequestCardsContainer;
